refactor: streamline token holder management and error handling in FastPullTokensService and FixHoldersService for improved data integrity and synchronization

+++ src/tokens/queues/sync-token-holders.queue.ts
@@ -51,16 +51,20 @@ export class SyncTokenHoldersQueue {
   async loadAndSaveTokenHoldersFromMdw(saleAddress: Encoded.ContractAddress) {
     const token = await this.tokenService.getToken(saleAddress);
     const aex9Address = await this.tokenService.getTokenAex9Address(token);
-    await this.tokenHoldersRepository.delete({
-      aex9_address: aex9Address,
-    });
+
     const totalHolders = await this.loadData(
       token,
       aex9Address,
       `${ACTIVE_NETWORK.middlewareUrl}/v3/aex9/${aex9Address}/balances?by=amount&limit=100`,
     );
+    if (totalHolders.length > 0) {
+      await this.tokenHoldersRepository.delete({
+        aex9_address: aex9Address,
+      });
+      await this.tokenHoldersRepository.upsert(totalHolders, ['id']);
+    }
     await this.tokensRepository.update(token.sale_address, {
-      holders_count: totalHolders,
+      holders_count: totalHolders.length,
     });
   }
 
@@ -68,7 +72,7 @@ export class SyncTokenHoldersQueue {
     token: Token,
     aex9Address: string,
     url: string,
-    totalHolders = 0,
+    totalHolders = [],
   ) {
     const response = await fetchJson(url);
     if (!response.data) {
@@ -84,6 +88,7 @@ export class SyncTokenHoldersQueue {
           .then((res) => {
             return Array.from(res)
               .map(([key, value]: any) => ({
+                id: `${key}_${aex9Address}`,
                 aex9_address: aex9Address,
                 address: key,
                 balance: new BigNumber(value),
@@ -91,8 +96,7 @@ export class SyncTokenHoldersQueue {
               .filter((item) => item.balance.gt(0))
               .sort((a, b) => b.balance.minus(a.balance).toNumber());
           });
-        await this.tokenHoldersRepository.save(holders);
-        return holders.length;
+        return holders || [];
       }
       this.logger.error(
         `SyncTokenHoldersQueue:failed to load data from url::${url}`,
@@ -113,7 +117,8 @@ export class SyncTokenHoldersQueue {
             holderData,
           );
         }
-        await this.tokenHoldersRepository.save({
+        totalHolders.push({
+          id: `${holderData?.account || holder.account_id}_${aex9Address}`,
           aex9_address: aex9Address,
           address: holderData?.account || holder.account_id,
           balance: new BigNumber(holderData?.amount || 0),
@@ -132,10 +137,10 @@ export class SyncTokenHoldersQueue {
         token,
         aex9Address,
         `${ACTIVE_NETWORK.middlewareUrl}${response.next}`,
-        totalHolders + holders.length,
+        totalHolders,
       );
     }
 
-    return totalHolders + holders.length;
+    return totalHolders;
   }
 }
